+++ modules/task/execution.go
@@ -65,8 +65,7 @@ type fnSignature struct {
 
 // Execute executes the function
 func (s *fnSignature) Execute(ctx context.Context) ([]reflect.Value, error) {
-	// TODO(madhu): Add back after module refactor
-	//globalBackendStatsClient().TaskExecutionCount().Inc(1)
+	globalBackendStatsClient().TaskExecutionCount().Inc(1)
 	targetArgs := make([]reflect.Value, 0, len(s.Args)+1)
 	targetArgs = append(targetArgs, reflect.ValueOf(ctx))
 	for _, arg := range s.Args {
@@ -85,15 +84,13 @@ func Enqueue(fn interface{}, args ...interface{}) error {
 	fnName := getFunctionName(fn)
 	_, ok := fnLookup.getFn(fnName)
 	if !ok {
-		// TODO(madhu): Add back after module refactor
-		//globalBackendStatsClient().TaskPublishFail().Inc(1)
+		globalBackendStatsClient().TaskPublishFail().Inc(1)
 		return fmt.Errorf("function: %q not found. Did you forget to register?", fnName)
 	}
 	fnType := reflect.TypeOf(fn)
 	// Validate function against arguments
 	if err := validateFnAgainstArgs(fnType, args); err != nil {
-		// TODO(madhu): Add back after module refactor
-		//globalBackendStatsClient().TaskPublishFail().Inc(1)
+		globalBackendStatsClient().TaskPublishFail().Inc(1)
 		return err
 	}
 	// Publish function to the backend
@@ -101,12 +98,10 @@ func Enqueue(fn interface{}, args ...interface{}) error {
 	s := fnSignature{FnName: fnName, Args: args[1:]}
 	sBytes, err := GlobalBackend().Encoder().Marshal(s)
 	if err != nil {
-		// TODO(madhu): Add back after module refactor
-		//globalBackendStatsClient().TaskPublishFail().Inc(1)
+		globalBackendStatsClient().TaskPublishFail().Inc(1)
 		return errors.Wrap(err, "unable to encode the function or args")
 	}
-	// TODO(madhu): Add back after module refactor
-	//globalBackendStatsClient().TaskPublishCount().Inc(1)
+	globalBackendStatsClient().TaskPublishCount().Inc(1)
 	return GlobalBackend().Enqueue(ctx, sBytes)
 }
 
@@ -149,9 +144,8 @@ func MustRegister(fn interface{}) {
 
 // Run decodes the message and executes as a task
 func Run(ctx context.Context, message []byte) error {
-	// TODO(madhu): Add back after module refactor
-	//stopwatch := globalBackendStatsClient().TaskExecutionTime().Start()
-	//defer stopwatch.Stop()
+	stopwatch := globalBackendStatsClient().TaskExecutionTime().Start()
+	defer stopwatch.Stop()
 
 	var s fnSignature
 	if err := GlobalBackend().Encoder().Unmarshal(message, &s); err != nil {
@@ -160,8 +154,7 @@ func Run(ctx context.Context, message []byte) error {
 	// TODO (madhu): Do we need a timeout here?
 	retValues, err := s.Execute(ctx)
 	if err != nil {
-		// TODO(madhu): Add back after module refactor
-		//globalBackendStatsClient().TaskExecuteFail().Inc(1)
+		globalBackendStatsClient().TaskExecuteFail().Inc(1)
 		return err
 	}
 	// Assume only an error will be returned since that is verified before adding to fnRegister